requirements.txt
confluent-kafka==2.3.0
psycopg2-binary==2.9.9
pandas==2.1.4
minio==7.2.3
faker==24.0.0
marimo==0.9.14
matplotlib==3.8.2
plotly==5.18.0
